+++ assets/javascripts/kitten/index.js
@@ -263,6 +263,7 @@ export { ButtonIcon } from './components/molecules/buttons/button-icon'
 export { ButtonGroup } from './components/molecules/buttons/button-group'
 export { ButtonImage } from './components/molecules/buttons/button-image'
 export { ButtonQuestionMarkIcon } from './components/molecules/buttons/button-question-mark-icon'
+export { ButtonWithTooltip } from './components/molecules/buttons/button-with-tooltip'
 export { CheckableButton } from './components/molecules/buttons/checkable-button'
 export { CloseButton } from './components/molecules/buttons/close-button'
 export {