+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `ButtonWithTooltip`: Add new component.
 - Feature: `DashboardLayout`: Add `Toaster` subcomponent.
 
 ## [4.9.0] - 2021-09-24
@@ -58,7 +59,7 @@ Fix:
 
 ## [4.6.0] - 2021-09-09
 
-Tech: 
+Tech:
 - 
 - Bump to `node-sass@6`.